backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
twilio==9.4.5
python-dotenv==1.0.1
email-validator==2.2.0
bcrypt==4.0.1